* Creation of setup.py and __init__.py files * Examples folder renamed to tests * Renamed blockdiagrams.py to diagram_builder.py
