+++ blockdiagrams/__init__.py
@@ -1,0 +1,3 @@
+from .diagram_builder import DiagramBuilder
+
+__all__ = ["DiagramBuilder"]
