+++ tests/block_angled.png
