+++ tests/block_rotation.png
